feat(ARC-177): Add solutions for problems A and B
This commit introduces solutions for problems A and B in ARC-177.
In problem A, we're given a set of coins and a list of amounts. The
solution checks if it's possible to make each amount using the coins.

In problem B, we're given a string of 1s and the solution generates a
new string based on the position of each 1 in the input string. The
length of the generated string and the string itself are printed.